api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ilya Vikulin <hi@ivikulin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.10"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
